chore: reorganize dataset internals
