+++ src/mvp_dataset/core/stages/map.py
@@ -1,0 +1,28 @@
+"""Map stage."""
+
+from __future__ import annotations
+
+from collections.abc import Callable, Iterable
+from dataclasses import dataclass
+
+from ..resume import ResumeStateError, callable_fingerprint, stable_fingerprint
+
+
+@dataclass(frozen=True, slots=True)
+class _MapStage:
+    fn: Callable[[object], object]
+
+    def __call__(self, data: Iterable[object]) -> Iterable[object]:
+        for sample in data:
+            yield self.fn(sample)
+
+    def state_dict(self) -> dict[str, object]:
+        return {}
+
+    def load_state_dict(self, state: dict[str, object]) -> None:
+        if state != {}:
+            msg = "[InvalidResumeState] map stage state must be empty"
+            raise ResumeStateError(msg)
+
+    def fingerprint(self) -> str:
+        return stable_fingerprint({"kind": "map", "fn": callable_fingerprint(self.fn)})
