+++ src/mvp_dataset/sources/__init__.py
@@ -1,24 +1,8 @@
 """Data source implementations."""
 
-from .jsonl import iter_jsonls
-from .lance import LanceDatasetSpec, LanceSourceSpec, list_lance_sources
-from .parquet import (
-    ParquetFragment,
-    iter_parquet,
-    iter_parquets,
-    list_parquet_fragments,
-)
-from .tar import iter_tar, iter_tars
+from .jsonl.dataset import JsonlDataset
+from .lance.dataset import LanceDataset
+from .parquet.dataset import ParquetDataset
+from .tar.dataset import TarDataset
 
-__all__ = [
-    "LanceDatasetSpec",
-    "LanceSourceSpec",
-    "ParquetFragment",
-    "iter_tar",
-    "iter_tars",
-    "iter_jsonls",
-    "iter_parquet",
-    "iter_parquets",
-    "list_lance_sources",
-    "list_parquet_fragments",
-]
+__all__ = ["JsonlDataset", "LanceDataset", "ParquetDataset", "TarDataset"]
