+++ src/mvp_dataset/sources/jsonl/__init__.py
@@ -1,19 +1,5 @@
-from .utils import (
-    TarManager,
-    TarRef,
-    iter_jsonls,
-    iter_ref_field_uris,
-    materialize_jsonl_shards,
-    parse_tar_uri,
-    resolve_ref_field_value,
-)
+"""JSONL source."""
 
-__all__ = [
-    "TarManager",
-    "TarRef",
-    "iter_jsonls",
-    "iter_ref_field_uris",
-    "materialize_jsonl_shards",
-    "parse_tar_uri",
-    "resolve_ref_field_value",
-]
+from .dataset import JsonlDataset
+
+__all__ = ["JsonlDataset"]
