+++ src/mvp_dataset/sources/jsonl/sharding.py
@@ -1,4 +1,4 @@
-"""JSONL source utilities, spill sharding, and tar reference parsing."""
+"""JSONL splitting and spill sharding."""
 
 from __future__ import annotations
 
@@ -6,16 +6,11 @@ import hashlib
 import json
 import os
 import subprocess
-import tarfile
-from collections import OrderedDict
-from collections.abc import Iterator, Sequence
-from dataclasses import dataclass
+from collections.abc import Sequence
 from pathlib import Path
-from types import TracebackType
 
-from ...core.types import PathLikeStr, RefFieldSpec, Sample
-
-_TAR_URI_PREFIX = "tar://"
+from ...core.types import PathLikeStr, Sample
+from .reader import _parse_jsonl_line
 
 
 def _wc_lines(path: str) -> int:
@@ -74,179 +69,6 @@ def split_jsonl_files(paths: list[str], min_chunks: int) -> list[str]:
         result_paths.extend(chunk_files)
 
     return result_paths
-
-
-@dataclass(frozen=True, slots=True)
-class TarRef:
-    """Parsed tar reference for one field in a JSONL sample."""
-
-    shard_path: str
-    key: str
-    field: str
-    raw_uri: str
-
-
-def parse_tar_uri(
-    uri: str,
-    *,
-    base_dir: PathLikeStr,
-    key_dot_level: int = 1,
-) -> TarRef:
-    """Parse URI with grammar ``tar://<shard>#<key>.<field>`` or ``<shard>#<key>.<field>``."""
-
-    body = uri[len(_TAR_URI_PREFIX) :] if uri.startswith(_TAR_URI_PREFIX) else uri
-    if "#" not in body:
-        msg = f"[InvalidTarUri] uri={uri!r} missing '#'"
-        raise ValueError(msg)
-
-    shard_part, target = body.split("#", 1)
-    if not shard_part:
-        msg = f"[InvalidTarUri] uri={uri!r} empty shard path"
-        raise ValueError(msg)
-    if key_dot_level <= 0:
-        msg = f"[InvalidTarUri] uri={uri!r} invalid key_dot_level={key_dot_level}"
-        raise ValueError(msg)
-    if "." not in target:
-        msg = f"[InvalidTarUri] uri={uri!r} expected '<key>.<field>' target"
-        raise ValueError(msg)
-
-    target_parts = target.split(".")
-    if len(target_parts) <= key_dot_level:
-        msg = f"[InvalidTarUri] uri={uri!r} target does not match key_dot_level={key_dot_level}"
-        raise ValueError(msg)
-
-    key = ".".join(target_parts[:key_dot_level])
-    field = ".".join(target_parts[key_dot_level:])
-    if not key or not field:
-        msg = f"[InvalidTarUri] uri={uri!r} invalid target key/field"
-        raise ValueError(msg)
-
-    shard_path = Path(shard_part)
-    if not shard_path.is_absolute():
-        shard_path = Path(base_dir) / shard_path
-    return TarRef(
-        shard_path=str(shard_path),
-        key=key,
-        field=field,
-        raw_uri=uri,
-    )
-
-
-def iter_ref_field_uris(value: object, *, field: str) -> Iterator[str]:
-    """Yield one-or-many tar reference URIs from a JSONL field value."""
-
-    if isinstance(value, str):
-        yield value
-        return
-    if isinstance(value, list):
-        for index, item in enumerate(value):
-            if not isinstance(item, str):
-                msg = (
-                    f"[InvalidRefField] expected string URI in field={field!r} list item "
-                    f"{index}, got={type(item).__name__}"
-                )
-                raise ValueError(msg)
-            yield item
-        return
-    msg = f"[InvalidRefField] expected string URI or list of string URIs in field={field!r} got={type(value).__name__}"
-    raise ValueError(msg)
-
-
-def resolve_ref_field_value(
-    value: object,
-    *,
-    field: str,
-    base_dir: PathLikeStr,
-    key_dot_level: int,
-    manager: TarManager,
-) -> object:
-    """Resolve one JSONL reference field to bytes or a list of bytes."""
-
-    if isinstance(value, str):
-        try:
-            tar_ref = parse_tar_uri(value, base_dir=base_dir, key_dot_level=key_dot_level)
-        except ValueError as exc:
-            msg = f"[InvalidRefField] failed to parse tar URI in field={field!r} value={value!r} reason={exc}"
-            raise ValueError(msg) from exc
-        return manager.read(tar_ref)
-
-    uris = list(iter_ref_field_uris(value, field=field))
-    resolved: list[bytes] = []
-    for uri in uris:
-        try:
-            tar_ref = parse_tar_uri(uri, base_dir=base_dir, key_dot_level=key_dot_level)
-        except ValueError as exc:
-            msg = f"[InvalidRefField] failed to parse tar URI in field={field!r} value={uri!r} reason={exc}"
-            raise ValueError(msg) from exc
-        resolved.append(manager.read(tar_ref))
-    return resolved
-
-
-class TarManager:
-    """LRU reader for tar-referenced field payloads."""
-
-    def __init__(self, max_open_files: int = 8) -> None:
-        if max_open_files < 1:
-            msg = f"[InvalidTarManagerSize] max_open_files must be >= 1, got={max_open_files}"
-            raise ValueError(msg)
-        self._entries: OrderedDict[str, tuple[tarfile.TarFile, dict[str, tarfile.TarInfo]]] = OrderedDict()
-        self._max_open_files = max_open_files
-
-    def __enter__(self) -> TarManager:
-        return self
-
-    def __exit__(
-        self,
-        _exc_type: type[BaseException] | None,
-        _exc: BaseException | None,
-        _tb: TracebackType | None,
-    ) -> None:
-        self.close()
-
-    def close(self) -> None:
-        for tf, _member_index in self._entries.values():
-            try:
-                tf.close()
-            except Exception:  # noqa: BLE001
-                pass
-        self._entries.clear()
-
-    def _get_tar_entry(
-        self,
-        shard_path: str,
-    ) -> tuple[tarfile.TarFile, dict[str, tarfile.TarInfo]]:
-        if shard_path in self._entries:
-            self._entries.move_to_end(shard_path)
-            return self._entries[shard_path]
-
-        if len(self._entries) >= self._max_open_files:
-            _evicted_path, (evicted_tf, _evicted_member_index) = self._entries.popitem(last=False)
-            try:
-                evicted_tf.close()
-            except Exception:  # noqa: BLE001
-                pass
-
-        tf = tarfile.open(shard_path, mode="r:*")
-        members = tf.getmembers()
-        member_index = {member.name: member for member in members}
-        entry = (tf, member_index)
-        self._entries[shard_path] = entry
-        return entry
-
-    def read(self, tar_ref: TarRef) -> bytes:
-        """Read and return the raw bytes payload described by *tar_ref*."""
-
-        member_name = f"{tar_ref.key}.{tar_ref.field}"
-        tf, member_index = self._get_tar_entry(tar_ref.shard_path)
-        member = member_index.get(member_name)
-        if member is None:
-            msg = f"[TarMemberNotFound] member={member_name!r} shard={tar_ref.shard_path!r} uri={tar_ref.raw_uri!r}"
-            raise KeyError(msg)
-        extracted = tf.extractfile(member)
-        if extracted is None:
-            msg = f"[TarExtractError] member={member_name!r} shard={tar_ref.shard_path!r} uri={tar_ref.raw_uri!r}"
-            raise tarfile.ExtractError(msg)
-        return extracted.read()
 
 
 def materialize_jsonl_shards(
@@ -354,36 +176,6 @@ def materialize_jsonl_shards(
     return [str(path) for path in shard_paths]
 
 
-def iter_jsonls(
-    shard_paths: Iterator[PathLikeStr],
-    ref_fields: tuple[RefFieldSpec, ...],
-) -> Iterator[Sample]:
-    """Resolve tar data references while streaming JSONL shard files."""
-    key_dot_level = int(os.environ.get("MVP_DATASET_TAR_KEY_DOT_LEVEL", "1"))
-    max_open_tars = int(os.environ.get("MVP_DATASET_TAR_MAX_OPEN_FILES", "8"))
-
-    def _resolve_one(sample: Sample, manager: TarManager) -> Sample:
-        resolved = dict(sample)
-        for field, base_dir in ref_fields:
-            if field not in sample:
-                continue
-            resolved[field] = resolve_ref_field_value(
-                sample[field],
-                field=field,
-                base_dir=base_dir,
-                key_dot_level=key_dot_level,
-                manager=manager,
-            )
-        return resolved
-
-    with TarManager(max_open_files=max_open_tars) as manager:
-        for shard_path in shard_paths:
-            with open(shard_path, encoding="utf-8") as handle:
-                for line_index, line in enumerate(handle):
-                    sample = _parse_jsonl_line(str(shard_path), line_index, line, allow_preannotated=True)
-                    yield _resolve_one(sample, manager)
-
-
 def _bucket_id_for_sample(sample: Sample, *, group_key: str | None, spill_buckets: int) -> int:
     if group_key is None:
         key = str(sample["__key__"])
@@ -423,40 +215,6 @@ def _jsonl_shard_plan_fingerprint(
     }
     encoded = json.dumps(payload, sort_keys=True, separators=(",", ":"), ensure_ascii=True).encode("utf-8")
     return hashlib.sha256(encoded).hexdigest()
-
-
-def _parse_jsonl_line(
-    file: str,
-    index_in_file: int,
-    line: str,
-    *,
-    allow_preannotated: bool = False,
-) -> Sample:
-    try:
-        parsed = json.loads(line)
-    except json.JSONDecodeError as exc:
-        msg = f"[InvalidJsonLine] file={file!r} line={index_in_file + 1} reason={exc.msg}"
-        raise ValueError(msg) from exc
-    if not isinstance(parsed, dict):
-        msg = f"[InvalidJsonSample] file={file!r} line={index_in_file + 1} expected object row"
-        raise ValueError(msg)
-
-    sample: Sample = dict(parsed)
-    if allow_preannotated and _has_jsonl_metadata(sample):
-        return sample
-
-    sample["__index_in_file__"] = index_in_file
-    sample["__file__"] = file
-    sample["__key__"] = f"{file}:{index_in_file}"
-    return sample
-
-
-def _has_jsonl_metadata(sample: Sample) -> bool:
-    return (
-        isinstance(sample.get("__index_in_file__"), int)
-        and isinstance(sample.get("__file__"), str)
-        and isinstance(sample.get("__key__"), str)
-    )
 
 
 def _resolve_final_shard_count(
