+++ src/mvp_dataset/sources/jsonl/types.py
@@ -1,0 +1,7 @@
+"""JSONL source types."""
+
+from __future__ import annotations
+
+from typing import Literal
+
+JsonlShuffleMode = Literal["none", "shard_aware", "global"]
