+++ src/mvp_dataset/sources/lance/refs/__init__.py
@@ -1,0 +1,24 @@
+"""Lance reference-column helpers."""
+
+from __future__ import annotations
+
+from ..types import LanceRefIndexScope
+from .config import attach_lance_ref_columns, parse_lance_ref_columns
+from .index import prepare_ref_indexes
+from .resolve import (
+    LanceRefResolverAssembler,
+    LanceResolveRefFactory,
+    iter_lance_ref_resolver,
+    validate_ref_names,
+)
+
+__all__ = [
+    "LanceRefIndexScope",
+    "LanceRefResolverAssembler",
+    "LanceResolveRefFactory",
+    "attach_lance_ref_columns",
+    "iter_lance_ref_resolver",
+    "parse_lance_ref_columns",
+    "prepare_ref_indexes",
+    "validate_ref_names",
+]
