+++ src/mvp_dataset/sources/parquet/fragments.py
@@ -1,9 +1,9 @@
-"""Parquet source implementation for row-group-parallel sample iteration."""
+"""Parquet fragment discovery."""
 
 from __future__ import annotations
 
 import os
-from collections.abc import Iterator, Sequence
+from collections.abc import Sequence
 from concurrent.futures import ThreadPoolExecutor
 from dataclasses import dataclass
 from itertools import repeat
@@ -11,11 +11,9 @@ from typing import Final
 
 import pyarrow.parquet as pq
 
-from ...core.types import PathLikeStr, Sample
+from ...core.types import PathLikeStr
 
-_DEFAULT_BATCH_SIZE: Final[int] = 65536
 _MAX_METADATA_WORKERS: Final[int] = 16
-_BATCH_SIZE_ENV_VAR: Final[str] = "MVP_DATASET_PARQUET_BATCH_SIZE"
 
 
 @dataclass(frozen=True, slots=True)
@@ -55,23 +53,6 @@ def _validate_min_row_groups_per_fragment(value: int) -> int:
         msg = f"[InvalidParquetFragmentConfig] min_row_groups_per_fragment must be >= 1, got {value}"
         raise ValueError(msg)
     return value
-
-
-def resolve_parquet_batch_size(batch_size: int | None = None) -> int:
-    if batch_size is None:
-        raw_value = os.environ.get(_BATCH_SIZE_ENV_VAR)
-        if raw_value is None:
-            batch_size = _DEFAULT_BATCH_SIZE
-        else:
-            try:
-                batch_size = int(raw_value)
-            except ValueError as exc:
-                msg = f"[InvalidParquetBatchSize] {_BATCH_SIZE_ENV_VAR} must be an integer, got {raw_value!r}"
-                raise ValueError(msg) from exc
-    if batch_size <= 0:
-        msg = f"[InvalidParquetBatchSize] {_BATCH_SIZE_ENV_VAR} must be >= 1, got {batch_size}"
-        raise ValueError(msg)
-    return batch_size
 
 
 def _collect_parquet_fragments(
@@ -151,59 +132,3 @@ def _collect_parquet_fragments_for_shard(
         )
 
     return fragments
-
-
-def iter_parquet(
-    fragment: ParquetFragment,
-    *,
-    columns: Sequence[str] | None = None,
-    batch_size: int | None = None,
-    use_threads: bool = True,
-    row_group_index: int | None = None,
-) -> Iterator[Sample]:
-    """Iterate one parquet row-group fragment and yield one sample dict per row."""
-
-    resolved_batch_size = resolve_parquet_batch_size(batch_size)
-    parquet_file = pq.ParquetFile(fragment.path)
-    row_groups = list(fragment.row_groups)
-    index_in_file = fragment.row_offset
-    if row_group_index is not None:
-        row_groups = [fragment.row_groups[row_group_index]]
-        index_in_file = fragment.row_group_offsets[row_group_index]
-
-    for record_batch in parquet_file.iter_batches(
-        batch_size=resolved_batch_size,
-        row_groups=row_groups,
-        columns=columns,
-        use_threads=use_threads,
-    ):
-        column_names = record_batch.schema.names
-        columns_data = [record_batch.column(i) for i in range(record_batch.num_columns)]
-        for batch_row_index in range(record_batch.num_rows):
-            sample: Sample = {
-                name: columns_data[column_index][batch_row_index].as_py()
-                for column_index, name in enumerate(column_names)
-            }
-            sample["__file__"] = fragment.path
-            sample["__index_in_file__"] = index_in_file
-            sample["__key__"] = f"{fragment.path}:{index_in_file}"
-            yield sample
-            index_in_file += 1
-
-
-def iter_parquets(
-    fragments: Iterator[ParquetFragment],
-    *,
-    columns: Sequence[str] | None = None,
-    batch_size: int | None = None,
-    use_threads: bool = True,
-) -> Iterator[Sample]:
-    """Iterate parquet row-group fragments in order and yield row samples."""
-
-    for fragment in fragments:
-        yield from iter_parquet(
-            fragment,
-            columns=columns,
-            batch_size=batch_size,
-            use_threads=use_threads,
-        )
