+++ src/mvp_dataset/sources/parquet/types.py
@@ -1,0 +1,7 @@
+"""Parquet source types."""
+
+from __future__ import annotations
+
+from typing import Literal
+
+ParquetShuffleMode = Literal["none", "fragment_aware", "global"]
