+++ src/mvp_dataset/sources/tar/types.py
@@ -1,0 +1,7 @@
+"""TAR source types."""
+
+from __future__ import annotations
+
+from typing import Literal
+
+TarShuffleMode = Literal["none", "shard_aware", "global"]
