+++ src/mvp_dataset/utils/__init__.py
@@ -1,6 +1,5 @@
-"""Utility helpers for path normalization and sharding."""
+"""Utility helpers."""
 
-from .sharding import assign_items
 from .url import normalize_paths
 
-__all__ = ["assign_items", "normalize_paths"]
+__all__ = ["normalize_paths"]
